bumboo-scraper/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
supabase>=2.4.0
python-dotenv>=1.0.0

